--- library/response.js
@@ -1,88 +1,0 @@
-class AssistantResponse extends AssistantResponseClass{
-  constructor (responseConfig, callbacks) {
-    super(responseConfig, callbacks)
-    this.hookChimed = false
-  }
-
-  start (response) {
-    super.start(response)
-    this.hookChimed = false
-  }
-
-  doCommand (commandName, param, from) {
-    if (!this.hookChimed) {
-      this.hookChimed = true
-      this.playChime("Beep")
-    }
-    super.doCommand(commandName, param, from)
-  }
-
-
-//You can implement these methods by your needs.
-/*
-  tunnel (payload) {
-    super.tunnel(payload)
-  }
-
-  doCommand (commandName, param, from) {
-    super.doCommand(commandName, param, from)
-  }
-
-  setSecret (secretMode) {
-    super.setSecret(secretMode)
-  }
-
-  playChime (sound) {
-    super.playChime(sound)
-  }
-
-  status (status) {
-    super.status(status)
-  }
-
-  prepare () {
-    super.prepare()
-  }
-
-  getDom () {
-    return super.getDom()
-  }
-
-  showError (text) {
-    super.showError(text)
-  }
-
-  showTranscription (text, className = "transcription") {
-    super.showTranscription(text, className)
-  }
-
-  end () {
-    super.end()
-  }
-
-  start (response) {
-    super.start(response)
-  }
-
-  stopResponse (callback = ()=>{}) {
-    super.stopResponse(callback)
-  }
-
-  postProcess (response, callback_done=()=>{}, callback_none=()=>{}) {
-    super.postProcess(response, callback_done, callback_none)
-  }
-
-  playAudioOutput (response) {
-    return super.playAudioOutput(response)
-  }
-
-  showScreenOutput (response) {
-    return super.showScreenOutput(response)
-  }
-
-  makeUrl (uri) {
-    return super.makeUrl(uri)
-  }
-
-*/
-}
